Правки функций
